fix ctc_train.py and move some common functions to common.py
